kms client operations
